bonds changes

+++ database/migrations/2023_03_15_031253_create_invoices_table.php
@@ -14,11 +14,14 @@ class CreateInvoicesTable extends Migration
     public function up()
     {
         Schema::create('invoices', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->integer('income_id');
+            $table->increments('id');
+
+            $table->unsignedBigInteger('income_id');
             $table->foreign('income_id')->references('id')->on('incomes');
-            $table->integer('course_id')->nullable();
+
+            $table->unsignedInteger('course_id')->nullable();
             $table->foreign('course_id')->references('id')->on('courses');
+
             $table->string('total');
             $table->string('remaining')->nullable();
             $table->string('user')->nullable();
