+++ database/migrations/2023_03_15_034845_create_bonds_table.php
@@ -15,22 +15,24 @@ class CreateBondsTable extends Migration
     {
         Schema::create('bonds', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('invoice_id');
+
+            $table->unsignedInteger('invoice_id');
             $table->foreign('invoice_id')->references('id')->on('invoices');
+            
             $table->integer('amount');
             
-            $table->integer('income_id');
+            $table->unsignedBigInteger('income_id');
             $table->foreign('income_id')->references('id')->on('incomes');
 
-            $table->integer('course_id')->nullable();
+            $table->unsignedInteger('course_id')->nullable();
             $table->foreign('course_id')->nullable()->references('id')->on('courses');
            
             $table->string('user');
 
-            $table->integer('createdBy');
+            $table->unsignedInteger('createdBy');
             $table->foreign('createdBy')->references('id')->on('users');
 
-            $table->integer('acceptedBy');
+            $table->unsignedInteger('acceptedBy');
             $table->foreign('acceptedBy')->references('id')->on('users');
             $table->timestamps();
         });
